11-modify-logic-to-allow-any-payer-from-the-expense-group feature added

+++ src/main/java/com/satwik/splitwiseclone/service/implementations/ExpenseServiceImpl.java
@@ -1,5 +1,6 @@
 package com.satwik.splitwiseclone.service.implementations;
 
+import com.satwik.splitwiseclone.exception.BadRequestException;
 import com.satwik.splitwiseclone.exception.DataNotFoundException;
 import com.satwik.splitwiseclone.persistence.dto.expense.*;
 import com.satwik.splitwiseclone.persistence.dto.user.OwerDTO;
@@ -30,6 +31,9 @@ public class ExpenseServiceImpl implements ExpenseService {
     private UserRepository userRepository;
 
     @Autowired
+    private GroupMembersRepository groupMembersRepository;
+
+    @Autowired
     private ExpenseShareRepository expenseShareRepository;
 
 
@@ -37,18 +41,24 @@ public class ExpenseServiceImpl implements ExpenseService {
     @Transactional
     public ExpenseDTO createGroupedExpense(UUID groupId, ExpenseDTO expenseDTO) {
 
-        User user = authorizationService.getAuthorizedUser();
+        User payer = expenseDTO.getPayerId() != null ? userRepository.findById(expenseDTO.getPayerId()).orElseThrow(() -> new DataNotFoundException("Payer not found!")) : authorizationService.getAuthorizedUser();
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new DataNotFoundException("Group not found"));
+
+        // checking if payer is a member of the group
+        if(!groupMembersRepository.existsByGroupIdAndMemberId(group.getId(), payer.getId()))
+            throw new BadRequestException("Payer is not a member of this group!");
+
         Expense expense = new Expense();
         expense.setAmount(expenseDTO.getAmount());
         expense.setDescription(expenseDTO.getDescription());
         expense.setGroup(group);
-        expense.setPayer(user);
+        expense.setPayer(payer);
         expenseRepository.save(expense);
 
         ExpenseDTO response = new ExpenseDTO();
         response.setExpenseId(expense.getId());
         response.setAmount(expense.getAmount());
+        response.setPayerId(expense.getPayer().getId());
         response.setDescription(expense.getDescription());
         response.setPayerName(expense.getPayer().getUsername());
         return response;
@@ -59,13 +69,14 @@ public class ExpenseServiceImpl implements ExpenseService {
     public ExpenseDTO createNonGroupedExpense(ExpenseDTO expenseDTO) {
 
         User user = authorizationService.getAuthorizedUser();
+        User payer = expenseDTO.getPayerId() != null ? userRepository.findById(expenseDTO.getPayerId()).orElseThrow(() -> new DataNotFoundException("Payer not found!")) : authorizationService.getAuthorizedUser();
         Group group = groupRepository.findDefaultGroup(user.getId()).orElseThrow(() -> new DataNotFoundException("Group not found"));
 
         Expense expense = new Expense();
         expense.setAmount(expenseDTO.getAmount());
         expense.setDescription(expenseDTO.getDescription());
         expense.setGroup(group);
-        expense.setPayer(user);
+        expense.setPayer(payer);
         expense = expenseRepository.save(expense);
 
         ExpenseDTO response = new ExpenseDTO();
@@ -73,6 +84,7 @@ public class ExpenseServiceImpl implements ExpenseService {
         response.setAmount(expense.getAmount());
         response.setDescription(expense.getDescription());
         response.setPayerName(expense.getPayer().getUsername());
+        response.setPayerId(expense.getPayer().getId());
 
         // TODO: add owner itself for
         return response;
@@ -91,11 +103,13 @@ public class ExpenseServiceImpl implements ExpenseService {
 
         Expense expense = expenseRepository.findById(expenseId).orElseThrow(() -> new RuntimeException("Expense not found"));
         User ower = userRepository.findById(owerId).orElseThrow(() -> new RuntimeException("Ower not found"));
+        if(expense.getPayer().getId().equals(ower.getId()))
+            throw new BadRequestException("Payer cannot be added as ower to the expense!");
         ExpenseShare expenseShare = new ExpenseShare();
         expenseShare.setExpense(expense);
         expenseShare.setUser(ower);
         double sharedAmount = expense.getAmount() /
-                (expenseShareRepository.findCountOfOwer(expenseId) + 1);
+                (expenseShareRepository.findCountOfOwer(expenseId) + 2); // old ower plus one for payer and one for new ower
         expenseShare.setSharedAmount(sharedAmount);
         List<ExpenseShare> shareList = expenseShareRepository.findExpenseShareById(expenseId);
         for (ExpenseShare share : shareList) {
@@ -111,8 +125,10 @@ public class ExpenseServiceImpl implements ExpenseService {
     public String removeUserFromExpense(UUID expenseId, UUID owerId) {
         Expense expense = expenseRepository.findById(expenseId).orElseThrow(() -> new DataNotFoundException("Expense not found"));
         User ower = userRepository.findById(owerId).orElseThrow(() -> new DataNotFoundException("Ower not found"));
+        if(!expenseShareRepository.existsByExpenseIdAndUserId(expense.getId(), ower.getId()))
+            throw new BadRequestException("Ower is not part of the expense. Hence, cannot be removed!");
         expenseShareRepository.deleteByExpenseIdAndUserId(expense.getId(), ower.getId());
-        double sharedAmount = expense.getAmount() / (expenseShareRepository.findCountOfOwer(expense.getId()));
+        double sharedAmount = expense.getAmount() / (expenseShareRepository.findCountOfOwer(expense.getId()) + 1);  // old ower plus one for payer
         List<ExpenseShare> shareList = expenseShareRepository.findExpenseShareById(expense.getId());
         for (ExpenseShare share : shareList) {
             share.setSharedAmount(sharedAmount);
@@ -131,6 +147,7 @@ public class ExpenseServiceImpl implements ExpenseService {
         expenseDTO.setDescription(expense.getDescription());
         expenseDTO.setAmount(expense.getAmount());
         expenseDTO.setPayerName(expense.getPayer().getUsername());
+        expenseDTO.setPayerId(expense.getPayer().getId());
         expenseDTO.setDate(expense.getCreatedOn());
         return expenseDTO;
     }
@@ -149,6 +166,7 @@ public class ExpenseServiceImpl implements ExpenseService {
             expenseDTO.setDate(expense.getCreatedOn());
             expenseDTO.setAmount(expense.getAmount());
             expenseDTO.setPayerName(expense.getPayer().getUsername());
+            expenseDTO.setPayerId(expense.getPayer().getId());
             expenseDTOS.add(expenseDTO);
         }
         return expenseDTOS;
